refactor: 컴포넌틑 추상화를 위한 디렉토리 구조 변경 및 GlobalStyles 변경
