+++ src/components/modal/Modal.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled, { keyframes } from 'styled-components';
-import Button from './Button';
+import Button from '../button/Button';
 
 const Modal = ({ type, onClose, children }) => {
   const ModalWrapper = SmallWrapper;
